+++ src/dbrequest/db_request.py
@@ -8,3 +8,4 @@ class DbRequest:
     db_type = DbType.NONE
     connection_string = None
     table_list:list = None
+    query = None